# swagger.yml
...
swagger: '2.0'
...
info:
  version: '1.0.0'
  title: Bi Bi Bi Bi Biotooooop

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
################################################################################
host: localhost
basePath: /api
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################
...
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: query
security:
  - api_key: []

################################################################################
#                                  Parameters                                  #
################################################################################
parameters:
...
  path_id:
    name: id
    in: path
    type: integer
    required: true
...
  body_User:
    name: body
    in: body
    schema:
      $ref: "#/definitions/UserStore"
...
paths:
...
  /store/users/{id}:
    get:
      tags:
        - store/users
      summary: Find the user by id
      parameters:
        - $ref: '#/parameters/path_id'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserStore"
        "404":
          description: Not found
    patch:
      tags:
        - store/users
      summary: Update the user sepcified by the id
      parameters:
        - $ref: '#/parameters/path_id'
        - $ref: '#/parameters/body_User'
      responses:
        "204":
          description: Success
        "400":
          description: Invalid input
        "404":
          description: Not found
    delete:
      tags:
        - store/users
      summary: Delete the user sepcified by the id
      parameters:
        - $ref: '#/parameters/path_id'
      responses:
        "204":
          description: Success
        "404":
          description: Not found
...
definitions:
  Token:
    type: object
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
  TokenStore:
    type: object
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      access_token_expires_in:
        type: string
        format: date-time
      refresh_token_expires_in:
        type: string
        format: date-time
  # TokenStore:
  #   type: object
  #   allOf:
  #   - $ref: '#/definitions/Token'
  #   - type: object
  #     properties:
  #       access_token_expires_in:
  #         type: string
  #         format: date-time
  #       refresh_token_expires_in:
  #         type: string
  #         format: date-time
  Tag:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      num_users:
        type: integer
        description: 登録ユーザ数
  Event:
    type: object
    properties:
      id:
        type: integer
      date:
        type: string
        format: date
      type:
        type: integer
        description: 1:縛りトープ / 2:ランダムトープ
        default: 1
        enum:
          - 1
          - 2
      num_users:
        type: integer
        description: エントリーユーザ数
  Biotop:
    type: object
    properties:
      id:
        type: integer
      event:
        $ref: "#/definitions/Event"
      tag:
        $ref: "#/definitions/Tag"
...
  User:
    type: object
    properties:
      id:
        type: integer
      email:
        type: string
        format: email
      name:
        type: string
      name_kana:
        type: string
      team:
        type: string
  UserStore:
    type: object
    properties:
      id:
        type: integer
      email:
        type: string
        format: email
      name:
        type: string
      name_kana:
        type: string
      team:
        type: string
      hitotalent_id:
        type: string
      gendar:
        type: string
      age:
        type: integer
      token:
        $ref: '#/definitions/TokenStore'
      tags:
        type: array
        items:
          $ref: "#/definitions/Tag"
      events:
        type: array
        items:
          $ref: "#/definitions/Event"
      biotops:
        type: array
        items:
          $ref: '#/definitions/UserBiotop'
  # UserStore:
  #   type: object
  #   allOf:
  #   - $ref: '#/definitions/User'
  #   - type: object
  #     properties:
  #       hitotalent_id:
  #         type: string
  #       gendar:
  #         type: string
  #       age:
  #         type: integer
  #       token:
  #         $ref: '#/definitions/TokenStore'
  #       tags:
  #         type: array
  #         items:
  #           $ref: "#/definitions/Tag"
  #       events:
  #         type: array
  #         items:
  #           $ref: "#/definitions/Event"
  #       biotops:
  #         type: array
  #         items:
  #           $ref: '#/definitions/UserBiotop'
  UserBiotop:
    type: object
    properties:
      id:
        type: integer
      event:
        $ref: "#/definitions/Event"
      tag:
        $ref: "#/definitions/Tag"
      canceled:
        type: boolean
        description: 確定後のキャンセル
  # UserBiotop:
  #   type: object
  #   allOf:
  #   - $ref: '#/definitions/Biotop'
  #   - $ref: '#/definitions/Canceled'
...